Enforce throttling on specific APIs. Add retry-after argument to 429 exception.
