+++ pyeasee/__init__.py
@@ -6,4 +6,5 @@ from .const import *  # noqa:
 from .easee import *  # noqa:
 from .easee import __VERSION__ as __version__  # noqa:
 from .site import *  # noqa:
+from .throttler import *  # noqa:
 from .utils import *  # noqa:
